+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 3.1.0
+
+- Added TrailAppearanceOverride API
+- Added CombatFlags API allowing to disable Better Combat on a per-player basis #600
+  - Supports disabling via java API call
+  - Supports disabling via entity tag command: `/tag @p add bettercombat_disabled`
+- Added `entity_scale_attack_range_influence` server config, to tune (or disable) how much entity scale affects attack range
+
 # 3.0.2
 
 - Refactor attack mechanic attachment (Interactor)